cast file read/write result to size_t

+++ src/file/file.h
@@ -39,8 +39,8 @@
 #define file_seek(X,Y,Z) X->seek(X,Y,Z)
 #define file_tell(X) X->tell(X)
 #define file_eof(X) X->eof(X)
-#define file_read(X,Y,Z) X->read(X,Y,Z)
-#define file_write(X,Y,Z) X->write(X,Y,Z)
+#define file_read(X,Y,Z) (size_t)X->read(X,Y,Z)
+#define file_write(X,Y,Z) (size_t)X->write(X,Y,Z)
 
 BD_PRIVATE extern BD_FILE_H* (*file_open)(const char* filename, const char *mode);
